requirements.txt
# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# AI and ML (simplified for initial deployment)
openai>=1.6.1,<2.0.0
langchain==0.0.340
langchain-openai==0.0.2

# Vector Database & RAG
chromadb==0.4.18
langchain-chroma==0.1.2
sentence-transformers==2.2.2
faiss-cpu==1.7.4
tiktoken==0.5.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Data Processing
pandas==2.1.3
numpy==1.25.2
requests==2.31.0
httpx==0.25.2

# Geospatial (simplified)
geopy==2.4.0

# Frontend & Real-time
streamlit==1.28.1
plotly==5.17.0
streamlit-autorefresh==0.0.1
streamlit-elements==0.1.0
websockets==12.0

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0

# Development
pytest==7.4.3
